feat: add test script
